tests/requirements.txt
alabaster==0.7.16
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
babel==2.16.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
commonmark==0.9.1
coverage==7.6.1
distlib==0.3.8
docopt==0.6.2
docutils==0.20.1
exceptiongroup==1.2.2
execnet==2.1.1
Faker==29.0.0
filelock==3.16.1
Flask==3.0.3
flit==3.9.0
flit_core==3.9.0
fluent-logger==0.11.1
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
grpc-interceptor==0.15.4
grpcio==1.66.1
grpcio-reflection==1.62.3
grpcio-status==1.62.3
grpcio-tools==1.70.0
httplib2==0.22.0
identify==2.6.1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.7
MarkupSafe==2.1.5
msgpack==1.1.0
nodeenv==1.9.1
packaging==24.1
paho-mqtt==1.6.1
pbr==6.1.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.8.0
proto-plus==1.24.0
protobuf==4.25.5
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
PyJWT==2.9.0
pykwalify==1.8.0
pyparsing==3.1.4
pyproject-api==1.8.0
pytest==7.2.2
pytest-cov==5.0.0
pytest-xdist==3.6.1
python-box==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
recommonmark==0.7.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.6.6
setuptools==75.8.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-markdown-tables==0.0.17
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stevedore==4.1.1
tavern==2.11.0
tomli==2.0.1
tomli_w==1.0.0
tox==4.20.0
types-jmespath==1.0.2.20240106
types-jsonschema==4.23.0.20240813
types-paho-mqtt==1.6.0.20240321
types-protobuf==4.25.0.20240417
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20240914
uritemplate==4.1.1
urllib3==2.2.3
uv==0.4.15
virtualenv==20.26.5
Werkzeug==3.0.4
wheel==0.44.0
zipp==3.20.2
